editor: use kakoune LSP
post-trying-out-using Kate when moving to KDE, one of the things I really
liked was having formatting and such managed by an external server that
is smarter as to the project structure than the editor. Language Server
Protocol daemons seem to fit this bill, and getting syntax errors in the
editor was pleasant too.